Minor rewrites to accommodate the new tensor library

+++ src/evolve/solver_trotter3.cc
@@ -19,7 +19,8 @@
 
 #include <mps/time_evolve.h>
 
-namespace mps {
+namespace mps
+{
 
   /**********************************************************************
    * Trotter's method with three passes
@@ -27,65 +28,83 @@ namespace mps {
    *	exp(-iHdt) = exp(-iH_even dt/2) exp(-iH_odd dt) exp(-iH_even dt/2)
    */
 
-  Trotter3Solver::Trotter3Solver(const Hamiltonian &H, cdouble dt) :
-  TrotterSolver(dt), U1(H, 1, dt), U2(H, 0, dt/2.0), sense(0)
+  Trotter3Solver::Trotter3Solver(const Hamiltonian &H, cdouble dt) : TrotterSolver(dt), U1(H, 1, dt), U2(H, 0, dt / 2.0), sense(0)
   {
   }
 
   double
   Trotter3Solver::one_step(CMPS *P, index Dmax)
   {
-    int debug = FLAGS.get(MPS_DEBUG_TROTTER);
-    if (!Dmax) {
-      if (strategy != DO_NOT_TRUNCATE) {
+    int debug = static_cast<int>(FLAGS.get(MPS_DEBUG_TROTTER));
+    if (!Dmax)
+    {
+      if (strategy != DO_NOT_TRUNCATE)
+      {
         std::cerr << "In TrotterSolver::one_step(), no maximum dimension provided\n";
         abort();
       }
     }
 
     U1.debug = U2.debug = debug;
-    switch (strategy) {
-    case TRUNCATE_EACH_UNITARY: {
+    switch (strategy)
+    {
+    case TRUNCATE_EACH_UNITARY:
+    {
       double err;
-      if (debug) std::cout << "Trotter3 method: truncate unitaries:\n"
-                           << "Trotter3 Layer 1/3\n";
+      if (debug)
+        std::cout << "Trotter3 method: truncate unitaries:\n"
+                  << "Trotter3 Layer 1/3\n";
       err = U2.apply(P, &sense, MPS_DEFAULT_TOLERANCE, Dmax);
-      if (debug) std::cout << "Trotter3 Layer 2/3\n";
+      if (debug)
+        std::cout << "Trotter3 Layer 2/3\n";
       err += U1.apply(P, &sense, MPS_DEFAULT_TOLERANCE, Dmax);
-      if (debug) std::cout << "Trotter3 Layer 3/3\n";
+      if (debug)
+        std::cout << "Trotter3 Layer 3/3\n";
       err += U2.apply(P, &sense, MPS_DEFAULT_TOLERANCE, Dmax, normalize);
       return err;
     }
-    case TRUNCATE_EACH_LAYER: {
+    case TRUNCATE_EACH_LAYER:
+    {
       double err = 0.0;
-      if (debug) std::cout << "Trotter3 method: truncate layers:\n"
-                           << "Trotter3 Layer 1/3\n";
+      if (debug)
+        std::cout << "Trotter3 method: truncate layers:\n"
+                  << "Trotter3 Layer 1/3\n";
       err = U2.apply_and_simplify(P, &sense, MPS_TRUNCATE_EPSILON, Dmax);
-      if (debug) std::cout << "Trotter3 Layer 2/3\n";
+      if (debug)
+        std::cout << "Trotter3 Layer 2/3\n";
       err += U1.apply_and_simplify(P, &sense, MPS_TRUNCATE_EPSILON, Dmax);
-      if (debug) std::cout << "Trotter3 Layer 3/3\n";
+      if (debug)
+        std::cout << "Trotter3 Layer 3/3\n";
       err += U2.apply_and_simplify(P, &sense, MPS_TRUNCATE_EPSILON, Dmax,
                                    normalize);
       return err;
     }
-    case DO_NOT_TRUNCATE: {
-      if (debug) std::cout << "Trotter3 method: do not truncate:\n"
-                           << "Trotter3 Layer 1/3\n";
+    case DO_NOT_TRUNCATE:
+    {
+      if (debug)
+        std::cout << "Trotter3 method: do not truncate:\n"
+                  << "Trotter3 Layer 1/3\n";
       U2.apply(P, &sense, MPS_TRUNCATE_EPSILON, 0);
-      if (debug) std::cout << "Trotter3 Layer 2/3\n";
+      if (debug)
+        std::cout << "Trotter3 Layer 2/3\n";
       U1.apply(P, &sense, MPS_TRUNCATE_EPSILON, 0);
-      if (debug) std::cout << "Trotter3 Layer 3/3\n";
+      if (debug)
+        std::cout << "Trotter3 Layer 3/3\n";
       U2.apply(P, &sense, MPS_TRUNCATE_EPSILON, 0, normalize);
       return 0.0;
     }
-    default: {
+    default:
+    {
       CMPS Pfull = *P;
-      if (debug) std::cout << "Trotter3 method: truncate group:\n"
-                           << "Trotter3 Layer 1/3\n";
+      if (debug)
+        std::cout << "Trotter3 method: truncate group:\n"
+                  << "Trotter3 Layer 1/3\n";
       U2.apply(P, &sense, MPS_TRUNCATE_EPSILON, 0);
-      if (debug) std::cout << "Trotter3 Layer 2/3\n";
+      if (debug)
+        std::cout << "Trotter3 Layer 2/3\n";
       U1.apply(P, &sense, MPS_TRUNCATE_EPSILON, 0);
-      if (debug) std::cout << "Trotter3 Layer 3/3\n";
+      if (debug)
+        std::cout << "Trotter3 Layer 3/3\n";
       return U2.apply_and_simplify(P, &sense, MPS_TRUNCATE_EPSILON, Dmax,
                                    normalize);
     }
